+++ mc2/controllers/base/tests/test_views.py
@@ -35,10 +35,17 @@ class ViewsTestCase(ControllerBaseTestCase):
             reverse('organizations:select-active', args=('foo-org',)))
 
         self.mock_create_marathon_app()
+        self.mock_create_postgres_db(200, {
+            'result': {
+                'name': 'joes_db',
+                'user': 'joe',
+                'password': '1234',
+                'host': 'localhost'}})
 
         data = {
             'name': 'Another test app',
             'marathon_cmd': 'ping2',
+            'postgres_db_needed': True,
             'env-TOTAL_FORMS': 0,
             'env-INITIAL_FORMS': 0,
             'env-MIN_NUM_FORMS': 0,
@@ -47,6 +54,12 @@ class ViewsTestCase(ControllerBaseTestCase):
             'label-INITIAL_FORMS': 0,
             'label-MIN_NUM_FORMS': 0,
             'label-MAX_NUM_FORMS': 100,
+            'link-TOTAL_FORMS': 0,
+            'link-INITIAL_FORMS': 0,
+            'link-MIN_NUM_FORMS': 0,
+            'link-MAX_NUM_FORMS': 100,
+
+
         }
 
         response = self.client.post(reverse('base:add'), data)
@@ -61,6 +74,62 @@ class ViewsTestCase(ControllerBaseTestCase):
         self.assertEqual(controller.marathon_cmd, 'ping2')
         self.assertEqual(controller.organization.slug, 'foo-org')
         self.assertTrue(controller.slug)
+        self.assertTrue(controller.postgres_db_needed)
+
+        self.assertEquals(controller.postgres_db_name, 'joes_db')
+        self.assertEquals(controller.postgres_db_username, 'joe')
+        self.assertEquals(controller.postgres_db_password, '1234')
+        self.assertEquals(controller.postgres_db_host, 'localhost')
+
+        resp = self.client.get(reverse('base:edit', args=[controller.id]))
+        self.assertContains(resp, 'postgres://joe:1234@localhost/joes_db')
+
+    @responses.activate
+    def test_postgres_db_needed_false(self):
+        existing_controller = self.mk_controller()
+
+        self.client.login(username='testuser2', password='test')
+        self.client.get(
+            reverse('organizations:select-active', args=('foo-org',)))
+
+        self.mock_create_marathon_app()
+        self.mock_create_postgres_db(200, {
+            'result': {
+                'name': 'joes_db',
+                'user': 'joe',
+                'password': '1234',
+                'host': 'localhost'}})
+
+        data = {
+            'name': 'Another test app',
+            'marathon_cmd': 'ping2',
+            'env-TOTAL_FORMS': 0,
+            'env-INITIAL_FORMS': 0,
+            'env-MIN_NUM_FORMS': 0,
+            'env-MAX_NUM_FORMS': 100,
+            'label-TOTAL_FORMS': 0,
+            'label-INITIAL_FORMS': 0,
+            'label-MIN_NUM_FORMS': 0,
+            'label-MAX_NUM_FORMS': 100,
+            'link-TOTAL_FORMS': 0,
+            'link-INITIAL_FORMS': 0,
+            'link-MIN_NUM_FORMS': 0,
+            'link-MAX_NUM_FORMS': 100,
+        }
+
+        response = self.client.post(reverse('base:add'), data)
+
+        self.assertEqual(response.status_code, 302)
+
+        controller = Controller.objects.exclude(
+            pk=existing_controller.pk).get()
+        self.assertEqual(controller.state, 'done')
+
+        self.assertEqual(controller.name, 'Another test app')
+        self.assertEqual(controller.marathon_cmd, 'ping2')
+        self.assertEqual(controller.organization.slug, 'foo-org')
+        self.assertTrue(controller.slug)
+        self.assertFalse(controller.postgres_db_needed)
 
     @responses.activate
     def test_create_new_controller_with_label(self):
@@ -71,6 +140,12 @@ class ViewsTestCase(ControllerBaseTestCase):
             reverse('organizations:select-active', args=('foo-org',)))
 
         self.mock_create_marathon_app()
+        self.mock_create_postgres_db(200, {
+            'result': {
+                'name': 'joes_db',
+                'user': 'joe',
+                'password': '1234',
+                'host': 'localhost'}})
 
         data = {
             'name': 'Another test app',
@@ -86,6 +161,10 @@ class ViewsTestCase(ControllerBaseTestCase):
             'label-INITIAL_FORMS': 0,
             'label-MIN_NUM_FORMS': 0,
             'label-MAX_NUM_FORMS': 100,
+            'link-TOTAL_FORMS': 0,
+            'link-INITIAL_FORMS': 0,
+            'link-MIN_NUM_FORMS': 0,
+            'link-MAX_NUM_FORMS': 100,
         }
 
         response = self.client.post(reverse('base:add'), data)
@@ -108,6 +187,60 @@ class ViewsTestCase(ControllerBaseTestCase):
         self.assertTrue(controller.slug)
 
     @responses.activate
+    def test_create_new_controller_with_additional_links(self):
+        existing_controller = self.mk_controller()
+
+        self.client.login(username='testuser2', password='test')
+        self.client.get(
+            reverse('organizations:select-active', args=('foo-org',)))
+
+        self.mock_create_marathon_app()
+        self.mock_create_postgres_db(200, {
+            'result': {
+                'name': 'joes_db',
+                'user': 'joe',
+                'password': '1234',
+                'host': 'localhost'}})
+
+        data = {
+            'name': 'Another test app',
+            'description': 'A really lovely little app',
+            'marathon_cmd': 'ping2',
+            'env-TOTAL_FORMS': 0,
+            'env-INITIAL_FORMS': 0,
+            'env-MIN_NUM_FORMS': 0,
+            'env-MAX_NUM_FORMS': 100,
+            'label-TOTAL_FORMS': 0,
+            'label-INITIAL_FORMS': 0,
+            'label-MIN_NUM_FORMS': 0,
+            'label-MAX_NUM_FORMS': 100,
+            'link-0-name': 'A_TEST_Name',
+            'link-0-link': 'testurl.com',
+            'link-TOTAL_FORMS': 1,
+            'link-INITIAL_FORMS': 0,
+            'link-MIN_NUM_FORMS': 0,
+            'link-MAX_NUM_FORMS': 100,
+        }
+
+        response = self.client.post(reverse('base:add'), data)
+        self.assertEqual(response.status_code, 302)
+
+        controller = Controller.objects.exclude(
+            pk=existing_controller.pk).get()
+        self.assertEqual(controller.state, 'done')
+
+        self.assertEqual(controller.name, 'Another test app')
+        self.assertEqual(controller.description, 'A really lovely little app')
+        self.assertEqual(controller.marathon_cmd, 'ping2')
+        self.assertEqual(controller.organization.slug, 'foo-org')
+        self.assertEqual(controller.additional_link.all().count(), 1)
+        self.assertEqual(
+            controller.additional_link.all()[0].name, 'A_TEST_Name')
+        self.assertEqual(
+            controller.additional_link.all()[0].link, 'testurl.com')
+        self.assertTrue(controller.slug)
+
+    @responses.activate
     def test_create_new_controller_with_env(self):
         existing_controller = self.mk_controller()
 
@@ -116,6 +249,12 @@ class ViewsTestCase(ControllerBaseTestCase):
             reverse('organizations:select-active', args=('foo-org',)))
 
         self.mock_create_marathon_app()
+        self.mock_create_postgres_db(200, {
+            'result': {
+                'name': 'joes_db',
+                'user': 'joe',
+                'password': '1234',
+                'host': 'localhost'}})
 
         webhook_token = str(uuid.uuid4())
 
@@ -132,6 +271,10 @@ class ViewsTestCase(ControllerBaseTestCase):
             'label-INITIAL_FORMS': 0,
             'label-MIN_NUM_FORMS': 0,
             'label-MAX_NUM_FORMS': 100,
+            'link-TOTAL_FORMS': 0,
+            'link-INITIAL_FORMS': 0,
+            'link-MIN_NUM_FORMS': 0,
+            'link-MAX_NUM_FORMS': 100,
             'webhook_token': webhook_token,
         }
 
@@ -166,6 +309,10 @@ class ViewsTestCase(ControllerBaseTestCase):
             'label-INITIAL_FORMS': 0,
             'label-MIN_NUM_FORMS': 0,
             'label-MAX_NUM_FORMS': 100,
+            'link-TOTAL_FORMS': 0,
+            'link-INITIAL_FORMS': 0,
+            'link-MIN_NUM_FORMS': 0,
+            'link-MAX_NUM_FORMS': 100,
         }
         response = self.client.post(reverse('base:add'), data)
 
@@ -183,6 +330,10 @@ class ViewsTestCase(ControllerBaseTestCase):
             'label-INITIAL_FORMS': 0,
             'label-MIN_NUM_FORMS': 0,
             'label-MAX_NUM_FORMS': 100,
+            'link-TOTAL_FORMS': 0,
+            'link-INITIAL_FORMS': 0,
+            'link-MIN_NUM_FORMS': 0,
+            'link-MAX_NUM_FORMS': 100,
         }
         response = self.client.post(reverse('base:add'), data)
 
@@ -226,6 +377,12 @@ class ViewsTestCase(ControllerBaseTestCase):
             reverse('organizations:select-active', args=('foo-org',)))
 
         self.mock_create_marathon_app()
+        self.mock_create_postgres_db(200, {
+            'result': {
+                'name': 'joes_db',
+                'user': 'joe',
+                'password': '1234',
+                'host': 'localhost'}})
 
         data = {
             'name': 'Another test app',
@@ -239,6 +396,10 @@ class ViewsTestCase(ControllerBaseTestCase):
             'label-INITIAL_FORMS': 0,
             'label-MIN_NUM_FORMS': 0,
             'label-MAX_NUM_FORMS': 100,
+            'link-TOTAL_FORMS': 0,
+            'link-INITIAL_FORMS': 0,
+            'link-MIN_NUM_FORMS': 0,
+            'link-MAX_NUM_FORMS': 100,
         }
 
         response = self.client.post(reverse('base:add'), data)
@@ -265,6 +426,10 @@ class ViewsTestCase(ControllerBaseTestCase):
                 'label-INITIAL_FORMS': 0,
                 'label-MIN_NUM_FORMS': 0,
                 'label-MAX_NUM_FORMS': 100,
+                'link-TOTAL_FORMS': 0,
+                'link-INITIAL_FORMS': 0,
+                'link-MIN_NUM_FORMS': 0,
+                'link-MAX_NUM_FORMS': 100,
                 'webhook_token': webhook_token,
             })
         controller = Controller.objects.get(pk=controller.id)
@@ -282,6 +447,12 @@ class ViewsTestCase(ControllerBaseTestCase):
             reverse('organizations:select-active', args=('foo-org',)))
 
         self.mock_create_marathon_app()
+        self.mock_create_postgres_db(200, {
+            'result': {
+                'name': 'joes_db',
+                'user': 'joe',
+                'password': '1234',
+                'host': 'localhost'}})
 
         data = {
             'name': 'Another test app',
@@ -294,6 +465,10 @@ class ViewsTestCase(ControllerBaseTestCase):
             'label-INITIAL_FORMS': 0,
             'label-MIN_NUM_FORMS': 0,
             'label-MAX_NUM_FORMS': 100,
+            'link-TOTAL_FORMS': 0,
+            'link-INITIAL_FORMS': 0,
+            'link-MIN_NUM_FORMS': 0,
+            'link-MAX_NUM_FORMS': 100,
         }
 
         response = self.client.post(reverse('base:add'), data)
@@ -317,6 +492,10 @@ class ViewsTestCase(ControllerBaseTestCase):
                 'label-INITIAL_FORMS': 0,
                 'label-MIN_NUM_FORMS': 0,
                 'label-MAX_NUM_FORMS': 100,
+                'link-TOTAL_FORMS': 0,
+                'link-INITIAL_FORMS': 0,
+                'link-MIN_NUM_FORMS': 0,
+                'link-MAX_NUM_FORMS': 100,
             })
         controller = Controller.objects.get(pk=controller.id)
         self.assertEqual(controller.marathon_cpus, 0.5)
@@ -343,7 +522,8 @@ class ViewsTestCase(ControllerBaseTestCase):
         self.client.post(
             reverse('user_settings'), {'settings_level': 'expert'})
         resp = self.client.get(reverse('home'))
-        self.assertContains(resp, 'You do not have permission to create')
+        self.assertNotContains(resp, 'edit')
+        self.assertContains(resp, 'view')
 
     def test_normal_user_who_is_org_admin_can_create_sites(self):
         resp = self.client.get(reverse('home'))
@@ -548,6 +728,12 @@ class ViewsTestCase(ControllerBaseTestCase):
             'owner': User.objects.get(pk=2)})
 
         self.mock_create_marathon_app()
+        self.mock_create_postgres_db(200, {
+            'result': {
+                'name': 'joes_db',
+                'user': 'joe',
+                'password': '1234',
+                'host': 'localhost'}})
         controller.get_builder().build()
 
         self.mock_exists_on_marathon(controller.app_id)
@@ -581,6 +767,12 @@ class ViewsTestCase(ControllerBaseTestCase):
             'owner': User.objects.get(pk=2)})
 
         self.mock_create_marathon_app()
+        self.mock_create_postgres_db(200, {
+            'result': {
+                'name': 'joes_db',
+                'user': 'joe',
+                'password': '1234',
+                'host': 'localhost'}})
         controller.get_builder().build()
 
         self.mock_exists_on_marathon(controller.app_id, 404)
@@ -723,6 +915,12 @@ class ViewsTestCase(ControllerBaseTestCase):
         self.client.get(
             reverse('organizations:select-active', args=('foo-org',)))
         self.mock_create_marathon_app()
+        self.mock_create_postgres_db(200, {
+            'result': {
+                'name': 'joes_db',
+                'user': 'joe',
+                'password': '1234',
+                'host': 'localhost'}})
 
         data = {
             'name': 'Another test app',
@@ -745,6 +943,12 @@ class ViewsTestCase(ControllerBaseTestCase):
             'label-INITIAL_FORMS': 0,
             'label-MIN_NUM_FORMS': 0,
             'label-MAX_NUM_FORMS': 100,
+            'link-0-name': 'A_TEST_Name',
+            'link-0-link': 'testurl.com',
+            'link-TOTAL_FORMS': 1,
+            'link-INITIAL_FORMS': 0,
+            'link-MIN_NUM_FORMS': 0,
+            'link-MAX_NUM_FORMS': 100,
         }
 
         response = self.client.post(reverse('base:add'), data)
@@ -765,6 +969,8 @@ class ViewsTestCase(ControllerBaseTestCase):
         self.assertContains(response, 'the label value')
         self.assertContains(response, 'A_TEST_LABEL2')
         self.assertContains(response, 'the label value2')
+        self.assertContains(response, 'A_TEST_Name')
+        self.assertContains(response, 'testurl.com')
 
     @responses.activate
     def test_cloning_a_controller_with_new_values(self):
@@ -772,6 +978,12 @@ class ViewsTestCase(ControllerBaseTestCase):
         self.client.get(
             reverse('organizations:select-active', args=('foo-org',)))
         self.mock_create_marathon_app()
+        self.mock_create_postgres_db(200, {
+            'result': {
+                'name': 'joes_db',
+                'user': 'joe',
+                'password': '1234',
+                'host': 'localhost'}})
 
         data = {
             'name': 'Another test app',
@@ -792,6 +1004,10 @@ class ViewsTestCase(ControllerBaseTestCase):
             'label-INITIAL_FORMS': 0,
             'label-MIN_NUM_FORMS': 0,
             'label-MAX_NUM_FORMS': 100,
+            'link-TOTAL_FORMS': 0,
+            'link-INITIAL_FORMS': 0,
+            'link-MIN_NUM_FORMS': 0,
+            'link-MAX_NUM_FORMS': 100,
         }
 
         response = self.client.post(reverse('base:add'), data)
@@ -821,6 +1037,10 @@ class ViewsTestCase(ControllerBaseTestCase):
                 'label-INITIAL_FORMS': 0,
                 'label-MIN_NUM_FORMS': 0,
                 'label-MAX_NUM_FORMS': 100,
+                'link-TOTAL_FORMS': 0,
+                'link-INITIAL_FORMS': 0,
+                'link-MIN_NUM_FORMS': 0,
+                'link-MAX_NUM_FORMS': 100,
 
             })
         self.assertEqual(response.status_code, 302)